backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
aiosqlite>=0.19.0
python-multipart>=0.0.6
pillow>=10.2.0
aiofiles>=23.2.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
